Updates for j00, j02, and j03
j00: Updated the template to include a "shift" attribute for the VecGroup class to move the vector's matrix that can be set for each individual VecGroup object.
j02 and j03: Updated to include the new template and adjust the animation to have the correct notation for the vectors and to move the matrices closer to the vectors. Also added a representation of the vectors to j02 to match j03.